requirements.txt
PyMySQL==0.6.3
blist==1.3.6
cassandra-driver==2.1.2
click==3.3
future==0.14.3
futures==2.2.0
lz4==0.7.0
six==1.8.0

